kafka/serde
