+++ src/v/kafka/reflection.h
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2020 Vectorized, Inc.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the file licenses/BSL.md
+ *
+ * As of the Change Date specified in that file, in accordance with
+ * the Business Source License, use of this software will be governed
+ * by the Apache License, Version 2.0
+ */
+
+#pragma once
+
+#include "kafka/types.h"
+#include "reflection/adl.h"
+
+/*
+ * TODO: bytes is on its way out in favor of iobuf. however its still lingering
+ * around in some types that we'd like to checkpoint to disk. therefore, this
+ * temporary hack serializes bytes as an iobuf so that we can avoid dealing with
+ * on-disk data compatibility when finally removing the last bit of bytes.
+ */
+namespace reflection {
+
+template<>
+struct adl<kafka::member_protocol> {
+    void to(iobuf& out, kafka::member_protocol p) {
+        iobuf md = bytes_to_iobuf(p.metadata);
+        reflection::serialize(out, p.name, md);
+    }
+
+    kafka::member_protocol from(iobuf_parser& in) {
+        return kafka::member_protocol{
+          .name = adl<kafka::protocol_name>{}.from(in),
+          .metadata = iobuf_to_bytes(adl<iobuf>{}.from(in)),
+        };
+    }
+};
+} // namespace reflection
