+++ src/v/kafka/server/group.cc
@@ -15,6 +15,7 @@
 #include "cluster/simple_batch_builder.h"
 #include "kafka/protocol/schemata/describe_groups_response.h"
 #include "kafka/protocol/sync_group.h"
+#include "kafka/reflection.h"
 #include "kafka/server/group_manager.h"
 #include "kafka/server/logger.h"
 #include "likely.h"
