+++ src/v/kafka/types.h
@@ -137,26 +137,3 @@ std::ostream& operator<<(std::ostream&, const member_protocol&);
 using assignments_type = std::unordered_map<member_id, bytes>;
 
 } // namespace kafka
-
-/*
- * TODO: bytes is on its way out in favor of iobuf. however its still lingering
- * around in some types that we'd like to checkpoint to disk. therefore, this
- * temporary hack serializes bytes as an iobuf so that we can avoid dealing with
- * on-disk data compatibility when finally removing the last bit of bytes.
- */
-namespace reflection {
-template<>
-struct adl<kafka::member_protocol> {
-    void to(iobuf& out, kafka::member_protocol p) {
-        iobuf md = bytes_to_iobuf(p.metadata);
-        reflection::serialize(out, p.name, md);
-    }
-
-    kafka::member_protocol from(iobuf_parser& in) {
-        return kafka::member_protocol{
-          .name = adl<kafka::protocol_name>{}.from(in),
-          .metadata = iobuf_to_bytes(adl<iobuf>{}.from(in)),
-        };
-    }
-};
-} // namespace reflection
